Create the application's form
Created the application's form.
This includes all the different sections of the CV.
It also includes optional sections and the ability to toggle them on and off.
Furthermore, created a notice to explain the application's function.
The aforementioned notice is shown on every visit.
NOTE: This commit doesn't include any input validation for form elements.