Validate form inputs
Added validation to all form inputs.
Now, when a user submit a form, they will see all issues with their submission.
Indications to errors are inserted in each section/segment and above the preview button.
NOTE: sections that are toggled off, DO NOT trigger errors.
Sections that are collapsed DO trigger errors (assuming the section is toggled on).