Add the ability to generate the user's resume
Used "React-pdf" to create the user's resume.
After the user validates the form they're able to preview and download their CV.
The contents of the CV are dynamic and change depending on the contents of the form.
NOTE: The size of the document is unrestricted, i.e., users may exceed 1 page.